+++ .gitignore
@@ -1,3 +1,6 @@
+#IDEA Files
+.idea/
+
 # Files and directories created by pub.
 .dart_tool/
 .packages
