+++ pubspec.yaml
@@ -12,3 +12,10 @@ dependencies:
   analyzer: ^5.12.0
   collection: ^1.17.2
   custom_lint_builder: ^0.5.0
+
+dev_dependencies:
+# These packages are mandatory for some of tests
+  flutter:
+    sdk: flutter
+  test: ^1.24.6
+
