+++ test/avoid_global_state_test.dart
